About Game
MOTIVATION

SUBWAY SURFER has been one of the most played game since almost a decade. We all have played it at least once in our childhood, and so we as well. As we got a dive into Software Engineering field, we were already inspired a lot by that game, hence we thought to take our inspiration to a next level and build our CUBRO. Our project is a single player strategy game emphasizing logical thinking and planning. People often stress over reaction time and time management, which usually takes precedence over fast action and character involvement. This game can help them improve their reaction time and improve their logical thinking.



INTRODUCTION

In the fast growing field of software engineering and development and even more rapidly growing sector of game development, the future is hard to predict. We are working with this game as our software project. We chose this type of work for doing better with development cycle, development period, graphics, scripting, adopting new technology and animation. The name of our project game is “’CUBRO”. In a game project, a game is much more than just its software. It has to provide content to become enjoyable. This has an important effect on the game project as a whole. The software part of the project is not the only one, and it must be considered in connection to all other parts: The environment of the game, game plays, the artwork, and so on. The game is to pass all obstacles without collision to achieve a high score. “CUBRO” is available for all windows devices.
 